Created redirect to longform url path
Note: localhost is hardcoded into the redirect. Otherwise, the redirect is up and working.

However, we will need to add a validation so that urls that do not begin with 'http://''
are not accepted. We should also give a visual cue to this effect in the view - for example
Original url: http://[textbox].