Long urls must be at least 6 characters long
Added our first activerecord validation. The validation is on the ruby
side of the ORM.

Question: I was expecting a more informative error message, like
"cannot insert short string into database" or something like that.
However, the database insertion failed silently: it did not show
anything on the console. The program failed later, when it tried
to go to the database and get a piece of data that wasn't there.